@@ -1,25 +1,29 @@
-import './App.css';
-import { Route, Switch } from 'react-router-dom';
-import LandingPage from './components/LandingPage/LandingPage'
-import React from 'react';
-import Home from './components/Home/Home';
-import NavBar from './components/NavBar/NavBar';
-import Detail from './components/CountryDetail/CountryDetail';
-import ActivityCreate from './components/ActivityCreate/activityCreate';
-import NotFound from './components/NotFound/NotFound';
+import { Route, Routes } from "react-router-dom";
+import MisPronosticos from "./components/MisPronosticos/MisPronosticos";
+import Headerlogin from "./components/NavBar/HeaderLogin";
+import Landing from "./components/Landing/Landing";
+import Resultados from "./components/Resultados/Resultados";
+
+//importacion de páginas
 
 function App() {
   return (
-    <div>
-      <Switch> 
-      <Route exact path={'/'} component={LandingPage} />    
-      <Route exact path={'/home'} component={Home} />
-      <Route exact path={'/home/country/:id'} component={Detail} />
-      <Route exact path={'/home/activity'} component={ActivityCreate} />
-      <Route component={NotFound} />
-      </Switch>
-      <Route path={'/home'} render={({match}) => <NavBar match={match}/>} />  
-    </div>
+    <>
+      <Headerlogin />
+      <Routes>
+        <Route path={"/"} element={<Landing />} />
+        <Route path={"/mispronosticos"} element={<MisPronosticos />} />
+        <Route path={"/resultados"} element={<Resultados />} />
+        {/* <Route path={"/pagolisto"} element={<PagoListo />} />
+        <Route path={"*"} element={<NotFound />} />
+        <Route path={"/easter"} element={<Easter />} />
+        <Route path={"/codenothere/alertas"} element={<ProtectedRoute component={Alertas}/>} />
+        <Route path={"/codenothere/usuarios"} element={<ProtectedRoute component={ListaUsuarios}/>} />
+        <Route path={"/codenothere/admins"} element={<ProtectedRoute component={AgregarAdmin}/>} />
+        <Route path={"/codenothere/mensajes"} element={<ProtectedRoute component={PreguntasDirectas}/>} />
+        <Route path={"/user/:sub"} element={<ProtectedRoute component={UserPage}/>} />  */}
+      </Routes>
+    </>
   );
 }
 
