@@ -18,15 +18,19 @@
 //                       `=---='
 //     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 const server = require('./src/app.js');
-const { getAllCountries } = require('./src/controllers/countriesControllers.js');
+const { populateGames } = require('./src/controllers/gamesControllers.js');
+const { populateTeams } = require('./src/controllers/teamsControllers.js');
+// const { getAllCountries } = require('./src/controllers/countriesControllers.js');
 const { conn } = require('./src/db.js');
 
 
 // Syncing all the models at once.
 // conn.sync({ force: true }).then(() => {
-  conn.sync({ force: false }).then(() => {
+  conn.sync({ force: true }).then(() => {
   server.listen(process.env.PORT, async () => {
     console.log('%s listening at 3001'); // eslint-disable-line no-console
-    await getAllCountries()
+    await populateTeams();
+    setTimeout(populateGames, 1000)
+
   });
 });
