@@ -1,8 +1,8 @@
 const { Router } = require('express');
-const countriesRoutes = require('./countriesRoutes')
-const countryRoute = require('./countryRoute')
-const filterRoutes = require('./filterRoutes')
-const activityRoute = require('./activityRoute');
+const usersRoutes = require('./usersRoutes')
+const userRoute = require('./userRoutes')
+const teamsRoutes = require('./teamsRoutes')
+const gamesRoute = require('./gamesRoutes');
 
 // Importar todos los routers;
 // Ejemplo: const authRouter = require('./auth.js');
@@ -12,9 +12,9 @@ const router = Router();
 // Configurar los routers
 // Ejemplo: router.use('/auth', authRouter);
 
-router.use("/countries", countriesRoutes)
-router.use("/country", countryRoute)
-router.use("/filters", filterRoutes)
-router.use("/activities", activityRoute)
+router.use("/users", usersRoutes)
+router.use("/user", userRoute)
+router.use("/teams", teamsRoutes)
+router.use("/games", gamesRoute)
 
 module.exports = router;
